Re-wrote  GetPeople added pagination
Had to rewirte the enitre getpeople method to get pagination to work together with the search function intact.

Also made code more readable as it's not duplicate code with an extra startswith method.

Now wedo all checks and check the result in the end. Pagination and search optional and doesn't affect the results