Cleaned up and fixed code relating to search
In GetPersonInterest removed the duplicate code and just applied search after taking out interests if a search was made. Same as with the ApplySearch method for GetInterests.

Also removed ApplySearch method as it was too complicated to make it universal for several ViewModels. The search in code in each method is just 5 lines. So I moved search back into GetInterests after simplifying it. And removed ApplySearch